refactoring as we go
