+++ src/applications/discharge-wizard/actions/index.js
@@ -1,9 +1,9 @@
 export const DW_UPDATE_FIELD = 'DW_UPDATE_FIELD';
 
-export function updateField(key, value) {
+export const updateField = (key, value) => {
   return {
     type: DW_UPDATE_FIELD,
     key,
     value,
   };
-}
+};
